Added HTTP 550: Permission denied
Implemented security check for HTTP 550

+++ php/delete.php
@@ -1,22 +1,26 @@
 <?php
 	session_start();
-	include_once("../access/database_functions.php");
-	if(isset($_REQUEST['id'])){
-		$id = $_REQUEST['id'];
+	if(isset($_SESSION['user'])){
+		include_once("../access/database_functions.php");
+		if(isset($_REQUEST['id'])){
+			$id = $_REQUEST['id'];
 
-		//Stores image path to delete after database call
-		$post = get_post($id);
-		$image = $post['image'];
+			//Stores image path to delete after database call
+			$post = get_post($id);
+			$image = $post['image'];
 
-		//Database call
-		delete($id);
+			//Database call
+			delete($id);
 
-		//Deletes image if an image exists
-		if($image != "null"){
-			unlink($image);
-		}	
-		//Message to display
-		echo "Blogginnlegget har blitt slettet";
+			//Deletes image if an image exists
+			if($image != "null"){
+				unlink($image);
+			}	
+			//Message to display
+			echo "Blogginnlegget har blitt slettet";
+		}
+	}else{
+		echo "HTTP 550: Permission denied";
 	}
 
 ?>
